refactoring util.js
