+++ src/mock/profile.js
@@ -1,4 +1,4 @@
-import { getRandomInteger } from '../util';
+import { getRandomInteger } from '../utils/common';
 
 const getRandomRating = () => {
   const ratings = [
